practical_applications/vr_video_converter/requirements.txt
cupy-cuda12x>=12.0.0
PyNvVideoCodec>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
flask>=2.3.0
